requirements.txt
aiortc==1.6.0
aiohttp==3.9.0
websockets==11.0.3
opencv-python-headless==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
uvloop==0.19.0
python-socketio==5.10.0
qrcode[pil]==7.4.2
psutil==5.9.6
asyncio-mqtt==0.16.1
python-dotenv==1.0.0